[fix] 댓글 무한렌더링 수정
